+++ src/components/lists-and-tables/list-item/list-item.i18n.ts
@@ -1,0 +1,5 @@
+export const nlddListItemTranslations = {
+	'components.list-item.opens-in-new-tab-label': 'Opent in nieuw tabblad',
+};
+
+export type NLDDListItemTranslations = typeof nlddListItemTranslations;